refactor: atoms 폴더 생성 후 컴포넌트 이동 및 스토리북 수정
