+++ src/components/atoms/ProfileIcon/ProfileIcon.style.ts
