+++ src/components/common/InputsUserInfo/InputProfileImage/InputProfileImage.tsx
@@ -6,8 +6,8 @@ import React, { useCallback, useState } from 'react';
 import { useDropzone } from 'react-dropzone';
 import defaultProfileImage from '../../../../assets/images/defaultProfile.png';
 import { Theme } from '../../../../styles/theme';
-import Button from '../../Button/Button';
-import Label from '../../Label/Label';
+import Button from '../../../atoms/Button/Button';
+import Label from '../../../atoms/Label/Label';
 import ProfileImage from '../../Profile/ProfileImage/ProfileImage';
 import {
   profileImageSelectContainer,
