+++ src/components/common/InputsUserInfo/InputPw.tsx
@@ -3,7 +3,7 @@ import { INPUT_LIMIT } from '@/constants/input';
 import { css } from '@emotion/react';
 import React, { ChangeEvent, useEffect } from 'react';
 import { useCheckPassword } from '../../../hooks/common/inputsUserInfo/useCheckPassword';
-import Input from '../Input/Input';
+import Input from '../../atoms/Input/Input';
 
 interface InputPwProps {
   value: string;
