+++ src/pages/MyPage/HistoryPage/TabPage/BookmarksPage/BookmarksPage.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
-import PageNavigation from '@/components/common/Pagination/Pagination';
-import Button from '@/components/common/Button/Button';
+import PageNavigation from '@/components/atoms/Pagination/Pagination';
+import Button from '@/components/atoms/Button/Button';
 import CheckBox from '@/components/design/CheckBox/CheckBox';
 import List from '@/components/mypage/List/List';
 import ItemMyBookmarksPosts from '@/components/mypage/ListItem/ItemMyBookmarksPosts';
