+++ src/pages/MyPage/UpdatePage/UpdatePage.tsx
@@ -8,9 +8,9 @@ import InputNickname from '@/components/common/InputsUserInfo/InputNickname';
 import InputProfileImage from '@/components/common/InputsUserInfo/InputProfileImage/InputProfileImage';
 import InputPw from '@/components/common/InputsUserInfo/InputPw';
 import InputPwCheck from '@/components/common/InputsUserInfo/InputPwCheck';
-import Button from '@/components/common/Button/Button';
+import Button from '@/components/atoms/Button/Button';
 import Heading from '@/components/common/Heading/Heading';
-import Input from '@/components/common/Input/Input';
+import Input from '@/components/atoms/Input/Input';
 import { useMemberUpdate } from '@/hooks/mypage/memberUpdate/useMemberUpdate';
 import {
   btnContainer,
