+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../../components/common/Button/Button';
+import Button from '../../components/atoms/Button/Button';
 import Heading from '../../components/common/Heading/Heading';
 import InputCode from '../../components/common/InputsUserInfo/InputCode';
 import InputEmail from '../../components/common/InputsUserInfo/InputEmail';
